requirements.txt
pandas
numpy
matplotlib
seaborn
statsmodels
scikit-learn
pmdarima
streamlit
pyyaml
requests
scipy
torch
